requirements.txt
python-dotenv
numpy
pandas
requests
Jinja2
ruff
premailer
